renamed dm_browser to fit folder naming convention
