+++ frontend/zc_messaging/src/pages/index.js
@@ -1,4 +1,4 @@
-import DmBrowser from "./dm_browser/DmBrowser"
+import DmBrowser from "./dm-browser/DmBrowser"
 import MessagingBoard from "./message-board"
 import ChannelBrowser from "./channel-browser"
 import Threads from "./threads"
